+++ source/app/service-providers/links/index.ts
@@ -15,7 +15,8 @@
 import { ipcMain } from 'electron'
 import broadcastIpcMessage from '@common/util/broadcast-ipc-message'
 import ProviderContract from '../provider-contract'
-import AppServiceContainer from 'source/app/app-service-container'
+import FSAL from '@providers/fsal'
+import LogProvider from '@providers/log'
 
 /**
  * This class manages the coloured tags of the app. It reads the tags on each
@@ -23,16 +24,16 @@ import AppServiceContainer from 'source/app/app-service-container'
  */
 export default class LinkProvider extends ProviderContract {
   private readonly _fileLinkDatabase: Map<string, string[]>
-  private readonly _idLinkDatabase: Map<string, string[]>
+  private _fsalHistoryTimestamp: number
 
   /**
    * Create the instance on program start and initially load the tags.
    */
-  constructor (private readonly _app: AppServiceContainer) {
+  constructor (private readonly _logger: LogProvider, private readonly _fsal: FSAL) {
     super()
 
     this._fileLinkDatabase = new Map()
-    this._idLinkDatabase = new Map()
+    this._fsalHistoryTimestamp = 0
     // TODO: Add a set of duplicate IDs so we can inform the user so they can
     // fix this
 
@@ -51,6 +52,89 @@ export default class LinkProvider extends ProviderContract {
         return Object.fromEntries(this._fileLinkDatabase)
       }
     })
+
+    // // Listen to state changes within the FSAL
+    // this._app.fsal.on('fsal-state-changed', (which) => {
+    //   if (which === 'reset-history') { // 'fsal-state-changed', 'reset-history'
+    //     this._fsalHistoryTimestamp = 0
+    //   } else if (which === 'filetree') {
+    //     // Retrieve all new FSAL events and handle them
+    //     const events = this._app.fsal.filetreeHistorySince(this._fsalHistoryTimestamp)
+    //     for (const event of events) {
+    //       this._fsalHistoryTimestamp = event.timestamp
+    //       this._updateLinksFor(event.path)
+    //     }
+    //   }
+    // })
+  }
+
+  public async boot (): Promise<void> {
+    // Listen to state changes within the FSAL
+    this._fsal.on('fsal-state-changed', (which) => {
+      if (which === 'reset-history') { // 'fsal-state-changed', 'reset-history'
+        this._fsalHistoryTimestamp = 0
+      } else if (which === 'filetree') {
+        // Retrieve all new FSAL events and handle them
+        const events = this._fsal.filetreeHistorySince(this._fsalHistoryTimestamp)
+        for (const event of events) {
+          this._fsalHistoryTimestamp = event.timestamp
+          this._updateLinksFor(event.path, event.event)
+        }
+      }
+    })
+  }
+
+  /**
+   * Takes the string of a file and updates any links according to the descriptor
+   * retrieved from the FSAL
+   *
+   * @param   {FSALHistoryEvent}  event  The event to check
+   */
+  private _updateLinksFor (path: string, event: 'remove'|'add'|'change'): void {
+    if (event === 'remove') {
+      if (this._fileLinkDatabase.has(path)) {
+        this._fileLinkDatabase.delete(path)
+      }
+      return
+    }
+
+    // So here's the thing: We're simply passing down event paths from the FSAL.
+    // Normally, this is easy, but sometimes directories are changed. Since we
+    // do a lot of comparison and only update if absolutely certain we can add
+    // a little bit overhead here
+    const dirDescriptor = this._fsal.findDir(path)
+    if (dirDescriptor !== null && event === 'add') {
+      for (const child of dirDescriptor.children) {
+        this._updateLinksFor(child.path, event)
+      }
+      return
+    }
+
+    const descriptor = this._fsal.findFile(path)
+    if (descriptor === null || descriptor.type !== 'file') {
+      // File has likely been removed, or some other error
+      if (this._fileLinkDatabase.has(path)) {
+        this._fileLinkDatabase.delete(path)
+      }
+      return
+    }
+
+    const newLinks = descriptor.links
+    const oldLinks = this._fileLinkDatabase.get(path)
+
+    if (oldLinks === undefined) {
+      // New file reporting
+      this._fileLinkDatabase.set(path, newLinks)
+      broadcastIpcMessage('links')
+    } else {
+      const sameLinks = oldLinks.every((value, index) => value === newLinks[index])
+
+      if (!sameLinks) {
+        // Same file reporting different links
+        this._fileLinkDatabase.set(path, newLinks)
+        broadcastIpcMessage('links')
+      }
+    }
   }
 
   /**
@@ -58,58 +142,7 @@ export default class LinkProvider extends ProviderContract {
    * @return {Boolean} Returns true after successful shutdown
    */
   async shutdown (): Promise<void> {
-    this._app.log.verbose('Link provider shutting down ...')
-  }
-
-  /**
-   * Adds an array of links from a specific file to the database. This
-   * function assumes sourceIDs to be unique, so in case of a duplicate, the
-   * later-loaded file overrides the earlier loaded one.
-   *
-   * @param   {string}            sourcePath     The full path to the source file
-   * @param   {string[]}          outboundLinks  A collection of links
-   * @param   {string|undefined}  sourceID       The ID of the source (if applicable)
-   */
-  report (sourcePath: string, outboundLinks: string[], sourceID?: string): void {
-    // NOTE: The FSAL by now defaults to an empty string instead of undefined
-    if (sourceID === '') {
-      sourceID = undefined
-    }
-
-    // NOTE: To anyone who comes here thinking that one might be able to
-    // optimize the graph building below further by resolving links immediately
-    // when they arrive here: That won't work because most of the time the
-    // target files won't be loaded when the source file's links arrive here.
-
-    this._fileLinkDatabase.set(sourcePath, outboundLinks)
-    if (sourceID !== undefined) {
-      this._idLinkDatabase.set(sourceID, outboundLinks)
-    }
-    broadcastIpcMessage('links')
-  }
-
-  /**
-   * Removes any outbound links emanating from the given file from the
-   * database. This function assumes sourceIDs to be unique, so in case of
-   * a duplicate, removing any of these files will delete the links for all.
-   *
-   * @param   {string}            sourcePath     The full path to the source file
-   * @param   {string|undefined}  sourceID       The ID of the source (if applicable)
-   */
-  remove (sourcePath: string, sourceID?: string): void {
-    // NOTE: The FSAL by now defaults to an empty string instead of undefined
-    if (sourceID === '') {
-      sourceID = undefined
-    }
-
-    if (this._fileLinkDatabase.has(sourcePath)) {
-      this._fileLinkDatabase.delete(sourcePath)
-    }
-
-    if (sourceID !== undefined && this._idLinkDatabase.has(sourceID)) {
-      this._idLinkDatabase.delete(sourceID)
-    }
-    broadcastIpcMessage('links')
+    this._logger.verbose('Link provider shutting down ...')
   }
 
   /**